Change card edit menu order
