+++ lib/settings/widgets/tag_card.dart
@@ -42,8 +42,8 @@ class _TagCardState extends State<TagCard> {
               ),
             ),
             CardEditMenu(actions: [
-              getDeletePopupAction(context, widget.onPressDelete),
               getDuplicatePopupAction(context, widget.onPressDuplicate),
+              getDeletePopupAction(context, widget.onPressDelete),
             ]),
           ],
         ));
